pyproject.toml
[project]
name = "metahike-support-env"
version = "1.0.0"
description = "Customer Support AI Training Environment for OpenEnv"
authors = [
    { name = "Divyansh Dusad" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "openai",
    "pytest",
    "PyYAML",
    "openenv-core>=0.2.0"
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
